Effects Bundle: Scan for glsl effects
Now it's possible to scan for effects on several search paths.
Tests are done only for the default effects path. needs to create more effects paths.
It should be possible to override effects. Effects have a name and if there is another on a further search path with the same name, it should override the first.
Uses the .json config files of the shaders when scanning an effect. the scan method should be improve to search for the .frag and .vert files per .json file.
Needs to change the load method of the effects to allow for different base paths. Currently only using the data path.